backend/requirements.txt
# -------------------------------
# Core Backend
# -------------------------------
fastapi
uvicorn[standard]
sqlalchemy
alembic

# -------------------------------
# Validation & Settings
# -------------------------------
pydantic
pydantic-settings
python-dateutil
pydantic[email]
# -------------------------------
# File Uploads & Images
# -------------------------------
python-multipart
pillow

# -------------------------------
# QR Code Generation
# -------------------------------
qrcode[pil]

# -------------------------------
# Excel / CSV Handling
# -------------------------------
pandas
openpyxl
xlsxwriter

# -------------------------------
# Authentication & Security (future ready)
# -------------------------------
passlib[bcrypt]
python-jose[cryptography]

# -------------------------------
# Utilities
# -------------------------------
python-dotenv
loguru

# -------------------------------
# Email
# -------------------------------
email-validator
pytz

# -------------------------------
# Scheduling
# -------------------------------
apscheduler